+++ test/Microsoft.AspNetCore.Server.Kestrel.FunctionalTests/RequestTests.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Concurrent;
 using System.Globalization;
 using System.IO;
+using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Net.Sockets;
@@ -15,6 +16,8 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
+using Microsoft.AspNetCore.Server.Kestrel.Core;
+using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http;
 using Microsoft.AspNetCore.Server.Kestrel.Transport.Libuv.Internal.Networking;
 using Microsoft.AspNetCore.Testing;
 using Microsoft.AspNetCore.Testing.xunit;
@@ -34,6 +37,15 @@ namespace Microsoft.AspNetCore.Server.Kestrel.FunctionalTests
         private const int _connectionKeepAliveEventId = 9;
         private const int _connectionResetEventId = 19;
         private const int _semaphoreWaitTimeout = 2500;
+
+        public static TheoryData<ListenOptions> ConnectionAdapterData => new TheoryData<ListenOptions>
+        {
+            new ListenOptions(new IPEndPoint(IPAddress.Loopback, 0)),
+            new ListenOptions(new IPEndPoint(IPAddress.Loopback, 0))
+            {
+                ConnectionAdapters = { new PassThroughConnectionAdapter() }
+            }
+        };
 
         [Theory]
         [InlineData(10 * 1024 * 1024, true)]
@@ -630,6 +642,558 @@ namespace Microsoft.AspNetCore.Server.Kestrel.FunctionalTests
                         usedIds.Add(id);
                     }
                 }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task Http11KeptAliveByDefault(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            using (var server = new TestServer(TestApp.EchoAppChunked, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "GET / HTTP/1.1",
+                        "",
+                        "GET / HTTP/1.1",
+                        "Connection: close",
+                        "Content-Length: 7",
+                        "",
+                        "Goodbye");
+                    await connection.ReceiveEnd(
+                        "HTTP/1.1 200 OK",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 0",
+                        "",
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 7",
+                        "",
+                        "Goodbye");
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task Http10NotKeptAliveByDefault(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            using (var server = new TestServer(TestApp.EchoApp, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "GET / HTTP/1.0",
+                        "",
+                        "");
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 0",
+                        "",
+                        "");
+                }
+
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "POST / HTTP/1.0",
+                        "Content-Length: 11",
+                        "",
+                        "Hello World");
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "",
+                        "Hello World");
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task Http10KeepAlive(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            using (var server = new TestServer(TestApp.EchoAppChunked, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "GET / HTTP/1.0",
+                        "Connection: keep-alive",
+                        "",
+                        "POST / HTTP/1.0",
+                        "Content-Length: 7",
+                        "",
+                        "Goodbye");
+                    await connection.Receive(
+                        "HTTP/1.1 200 OK",
+                        "Connection: keep-alive",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 0",
+                        "\r\n");
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 7",
+                        "",
+                        "Goodbye");
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task Http10KeepAliveNotHonoredIfResponseContentLengthNotSet(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            using (var server = new TestServer(TestApp.EchoApp, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "GET / HTTP/1.0",
+                        "Connection: keep-alive",
+                        "",
+                        "");
+
+                    await connection.Receive(
+                        "HTTP/1.1 200 OK",
+                        "Connection: keep-alive",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 0",
+                        "\r\n");
+
+                    await connection.Send(
+                        "POST / HTTP/1.0",
+                        "Connection: keep-alive",
+                        "Content-Length: 7",
+                        "",
+                        "Goodbye");
+
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "",
+                        "Goodbye");
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task Http10KeepAliveHonoredIfResponseContentLengthSet(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            using (var server = new TestServer(TestApp.EchoAppChunked, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "POST / HTTP/1.0",
+                        "Content-Length: 11",
+                        "Connection: keep-alive",
+                        "",
+                        "Hello World");
+
+                    await connection.Receive(
+                        "HTTP/1.1 200 OK",
+                        "Connection: keep-alive",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 11",
+                        "",
+                        "Hello World");
+
+                    await connection.Send(
+                        "POST / HTTP/1.0",
+                        "Connection: keep-alive",
+                        "Content-Length: 11",
+                        "",
+                        "Hello Again");
+
+                    await connection.Receive(
+                        "HTTP/1.1 200 OK",
+                        "Connection: keep-alive",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 11",
+                        "",
+                        "Hello Again");
+
+                    await connection.Send(
+                        "POST / HTTP/1.0",
+                        "Content-Length: 7",
+                        "",
+                        "Goodbye");
+
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 7",
+                        "",
+                        "Goodbye");
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task Expect100ContinueHonored(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            using (var server = new TestServer(TestApp.EchoAppChunked, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "POST / HTTP/1.1",
+                        "Expect: 100-continue",
+                        "Connection: close",
+                        "Content-Length: 11",
+                        "\r\n");
+                    await connection.Receive(
+                        "HTTP/1.1 100 Continue",
+                        "",
+                        "");
+                    await connection.Send("Hello World");
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 11",
+                        "",
+                        "Hello World");
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task ZeroContentLengthAssumedOnNonKeepAliveRequestsWithoutContentLengthOrTransferEncodingHeader(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            using (var server = new TestServer(async httpContext =>
+            {
+                // This will hang if 0 content length is not assumed by the server
+                Assert.Equal(0, await httpContext.Request.Body.ReadAsync(new byte[1], 0, 1).TimeoutAfter(TimeSpan.FromSeconds(10)));
+            }, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    // Use Send instead of SendEnd to ensure the connection will remain open while
+                    // the app runs and reads 0 bytes from the body nonetheless. This checks that
+                    // https://github.com/aspnet/KestrelHttpServer/issues/1104 is not regressing.
+                    await connection.Send(
+                        "GET / HTTP/1.1",
+                        "Connection: close",
+                        "",
+                        "");
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 0",
+                        "",
+                        "");
+                }
+
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "GET / HTTP/1.0",
+                        "",
+                        "");
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 0",
+                        "",
+                        "");
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task ConnectionClosesWhenFinReceivedBeforeRequestCompletes(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            using (var server = new TestServer(TestApp.EchoAppChunked, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "POST / HTTP/1.1");
+                    connection.Shutdown(SocketShutdown.Send);
+                    await connection.ReceiveForcedEnd();
+                }
+
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "POST / HTTP/1.1",
+                        "Content-Length: 7");
+                    connection.Shutdown(SocketShutdown.Send);
+                    await connection.ReceiveForcedEnd();
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task RequestsCanBeAbortedMidRead(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            var readTcs = new TaskCompletionSource<object>();
+            var registrationTcs = new TaskCompletionSource<int>();
+            var requestId = 0;
+
+            using (var server = new TestServer(async httpContext =>
+            {
+                requestId++;
+
+                var response = httpContext.Response;
+                var request = httpContext.Request;
+                var lifetime = httpContext.Features.Get<IHttpRequestLifetimeFeature>();
+
+                lifetime.RequestAborted.Register(() => registrationTcs.TrySetResult(requestId));
+
+                if (requestId == 1)
+                {
+                    response.Headers["Content-Length"] = new[] { "5" };
+
+                    await response.WriteAsync("World");
+                }
+                else
+                {
+                    var readTask = request.Body.CopyToAsync(Stream.Null);
+
+                    lifetime.Abort();
+
+                    try
+                    {
+                        await readTask;
+                    }
+                    catch (Exception ex)
+                    {
+                        readTcs.SetException(ex);
+                        throw;
+                    }
+
+                    readTcs.SetException(new Exception("This shouldn't be reached."));
+                }
+            }, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    // Never send the body so CopyToAsync always fails.
+                    await connection.Send(
+                        "POST / HTTP/1.1",
+                        "Content-Length: 5",
+                        "",
+                        "HelloPOST / HTTP/1.1",
+                        "Content-Length: 5",
+                        "",
+                        "");
+
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 200 OK",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 5",
+                        "",
+                        "World");
+                }
+            }
+
+            await Assert.ThrowsAsync<TaskCanceledException>(async () => await readTcs.Task);
+
+            // The cancellation token for only the last request should be triggered.
+            var abortedRequestId = await registrationTcs.Task;
+            Assert.Equal(2, abortedRequestId);
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task RequestHeadersAreResetOnEachRequest(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+
+            IHeaderDictionary originalRequestHeaders = null;
+            var firstRequest = true;
+
+            using (var server = new TestServer(httpContext =>
+            {
+                var requestFeature = httpContext.Features.Get<IHttpRequestFeature>();
+
+                if (firstRequest)
+                {
+                    originalRequestHeaders = requestFeature.Headers;
+                    requestFeature.Headers = new FrameRequestHeaders();
+                    firstRequest = false;
+                }
+                else
+                {
+                    Assert.Same(originalRequestHeaders, requestFeature.Headers);
+                }
+
+                return TaskCache.CompletedTask;
+            }, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "GET / HTTP/1.1",
+                        "",
+                        "GET / HTTP/1.1",
+                        "",
+                        "");
+                    await connection.ReceiveEnd(
+                        "HTTP/1.1 200 OK",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 0",
+                        "",
+                        "HTTP/1.1 200 OK",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 0",
+                        "",
+                        "");
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task UpgradeRequestIsNotKeptAliveOrChunked(ListenOptions listenOptions)
+        {
+            const string message = "Hello World";
+
+            var testContext = new TestServiceContext();
+
+            using (var server = new TestServer(async context =>
+            {
+                var upgradeFeature = context.Features.Get<IHttpUpgradeFeature>();
+                var duplexStream = await upgradeFeature.UpgradeAsync();
+
+                var buffer = new byte[message.Length];
+                var read = 0;
+                while (read < message.Length)
+                {
+                    read += await duplexStream.ReadAsync(buffer, read, buffer.Length - read).TimeoutAfter(TimeSpan.FromSeconds(10));
+                }
+
+                await duplexStream.WriteAsync(buffer, 0, read);
+            }, testContext, listenOptions))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    await connection.Send(
+                        "GET / HTTP/1.1",
+                        "Connection: Upgrade",
+                        "",
+                        message);
+                    await connection.ReceiveForcedEnd(
+                        "HTTP/1.1 101 Switching Protocols",
+                        "Connection: Upgrade",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "",
+                        message);
+                }
+            }
+        }
+
+        [Theory]
+        [MemberData(nameof(ConnectionAdapterData))]
+        public async Task HeadersAndStreamsAreReusedAcrossRequests(ListenOptions listenOptions)
+        {
+            var testContext = new TestServiceContext();
+            var streamCount = 0;
+            var requestHeadersCount = 0;
+            var responseHeadersCount = 0;
+            var loopCount = 20;
+            Stream lastStream = null;
+            IHeaderDictionary lastRequestHeaders = null;
+            IHeaderDictionary lastResponseHeaders = null;
+
+            using (var server = new TestServer(async context =>
+            {
+                if (context.Request.Body != lastStream)
+                {
+                    lastStream = context.Request.Body;
+                    streamCount++;
+                }
+                if (context.Request.Headers != lastRequestHeaders)
+                {
+                    lastRequestHeaders = context.Request.Headers;
+                    requestHeadersCount++;
+                }
+                if (context.Response.Headers != lastResponseHeaders)
+                {
+                    lastResponseHeaders = context.Response.Headers;
+                    responseHeadersCount++;
+                }
+
+                var ms = new MemoryStream();
+                await context.Request.Body.CopyToAsync(ms);
+                var request = ms.ToArray();
+
+                context.Response.ContentLength = request.Length;
+
+                await context.Response.Body.WriteAsync(request, 0, request.Length);
+            }, testContext))
+            {
+                using (var connection = server.CreateConnection())
+                {
+                    var requestData =
+                        Enumerable.Repeat("GET / HTTP/1.1\r\n", loopCount)
+                            .Concat(new[] { "GET / HTTP/1.1\r\nContent-Length: 7\r\nConnection: close\r\n\r\nGoodbye" });
+
+                    var response = string.Join("\r\n", new string[] {
+                        "HTTP/1.1 200 OK",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 0",
+                        ""});
+
+                    var lastResponse = string.Join("\r\n", new string[]
+                    {
+                        "HTTP/1.1 200 OK",
+                        "Connection: close",
+                        $"Date: {testContext.DateHeaderValue}",
+                        "Content-Length: 7",
+                        "",
+                        "Goodbye"
+                    });
+
+                    var responseData =
+                        Enumerable.Repeat(response, loopCount)
+                            .Concat(new[] { lastResponse });
+
+                    await connection.Send(requestData.ToArray());
+
+                    await connection.ReceiveEnd(responseData.ToArray());
+                }
+
+                Assert.Equal(1, streamCount);
+                Assert.Equal(1, requestHeadersCount);
+                Assert.Equal(1, responseHeadersCount);
             }
         }
